feat: move branch and employee interfaces into new models directory
