+++ src/api/v1/models/branchModel.ts
@@ -1,0 +1,6 @@
+export interface Branch {
+    id: number,
+    name: string,
+    address: string,
+    phone: string
+}
