+++ src/api/v1/models/employeeModel.ts
@@ -1,0 +1,9 @@
+export interface Employee {
+    id: number,
+    name: string,
+    position: string,
+    department: string,
+    email: string,
+    phone: string,
+    branchId: number
+}
